refactor: refactored folders
