+++ src/components/accounts/SignUpBurger.tsx
@@ -2,7 +2,7 @@ import Modal from "react-modal"
 import { useState } from "react";
 import { useEffect } from "react";
 Modal.setAppElement('#root');
-import db from "../../firebase.config";
+import db from "../../../firebase.config";
 import { doc, setDoc, getDoc } from "firebase/firestore";
 
 const SignUpBurger = () => {
